requirements.txt
APScheduler==2.1.1
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
gunicorn==18.0
itsdangerous==0.23
psycopg2==2.5.1
pyserial==2.7
wsgiref==0.1.2
